server/python/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
faiss-cpu==1.7.0
filelock==3.0.12
gensim==3.8.3
idna==2.10
joblib==1.0.1
nltk==3.5
numpy==1.20.1
packaging==20.9
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
sentence-transformers==1.0.4
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
smart-open==4.2.0
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.8.1
tqdm==4.59.0
transformers==4.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
